fix: bind review rating as int16 to match INT2 column
Postgres rejected the prepared statement with "incorrect binary data
format in bind parameter 2" because review_rating is INT2 but the
backend bound a 32-bit int. Cast rating to short before binding.

+++ backend/src/controllers/ReviewController.cc
@@ -46,8 +46,8 @@ void ReviewController::create(const HttpRequestPtr& req,
     std::function<void(const HttpResponsePtr&)>&& cb){
     auto b=req->getJsonObject();
     if(!b) return cb(jsonErr("JSON required"));
-    int  clientId = (*b)["client_id"].asInt();
-    int  rating   = (*b)["rating"].asInt();
+    int   clientId = (*b)["client_id"].asInt();
+    short rating   = static_cast<short>((*b)["rating"].asInt());
     std::string comment = (*b).get("comment","").asString();
     // store_id опциональный
     bool hasStore = (*b).isMember("store_id") && !(*b)["store_id"].isNull()
